requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
blas=1.0=mkl
ca-certificates=2019.1.23=0
certifi=2018.11.29=py36_0
cffi=1.12.1=py36h2e261b9_0
cudatoolkit=9.0=h13b8566_0
cycler=0.10.0=py36_0
dbus=1.13.6=h746ee38_0
expat=2.2.6=he6710b0_0
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
glib=2.56.2=hd408876_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
icu=58.2=h9c2bf20_1
intel-openmp=2019.1=144
jpeg=9b=h024ee3a_2
kiwisolver=1.0.1=py36hf484d3e_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=8.2.0=hdf63c60_1
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.36=hbc83047_0
libstdcxx-ng=8.2.0=hdf63c60_1
libtiff=4.0.10=h2733197_2
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.9=he19cac6_0
matplotlib=3.0.2=py36h5429711_0
mkl=2019.1=144
mkl_fft=1.0.10=py36ha843d7b_0
mkl_random=1.0.2=py36hd81dba3_0
ncurses=6.1=he6710b0_1
ninja=1.8.2=py36h6bb024c_1
numpy=1.16.2=py36h7e9f1db_0
numpy-base=1.16.2=py36hde5b4d6_0
olefile=0.46=py36_0
openssl=1.1.1b=h7b6447c_1
pcre=8.42=h439df22_0
pillow=5.4.1=py36h34e0f95_0
pip=19.0.3=py36_0
pycparser=2.19=py36_0
pyparsing=2.3.1=py36_0
pyqt=5.9.2=py36h05f1152_2
python=3.6.8=h0371630_0
python-dateutil=2.8.0=py36_0
pytorch=1.0.1=py3.6_cuda9.0.176_cudnn7.4.2_2
pytz=2018.9=py36_0
qt=5.9.7=h5867ecd_1
readline=7.0=h7b6447c_5
setuptools=40.8.0=py36_0
sip=4.19.8=py36hf484d3e_0
six=1.12.0=py36_0
sqlite=3.26.0=h7b6447c_0
tk=8.6.8=hbc83047_0
torchvision=0.2.2=py_3
tornado=5.1.1=py36h7b6447c_0
wheel=0.33.1=py36_0
xz=5.2.4=h14c3975_4
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
